#19: починить сборку 0.4.26 — Long в coerceAtLeast (падение compileDebugKotlin)
Первая же сборка после правок упала на моей же строке из 48f6b11:
SpeechPlayer.kt:390 — SYNTH_STALL_MS объявлен Long (15_000L), поэтому
`SYNTH_STALL_MS / 1000` даёт Long, а coerceAtLeast у Int-значения ждёт Int.
Ловушка в том, что строку не с чем было сверить: локально компилятора нет
(arm64-контейнер), проверка — только сборкой на GitHub.

+++ app/src/main/java/com/cuber/bookvoice/SpeechPlayer.kt
@@ -387,7 +387,7 @@ class SpeechPlayer(context: Context) {
         // Реальное время молчания, а не номинал сторожа: если часы «убежали»
         // далеко за 15 с — значит процесс спал (msg4211, #19).
         val waited = ((SystemClock.elapsedRealtime() - watchFromMs) / 1000).toInt()
-            .coerceAtLeast(SYNTH_STALL_MS / 1000)
+            .coerceAtLeast((SYNTH_STALL_MS / 1000).toInt())
         stuck.forEach { (id, p) ->
             p.cancelled = true
             p.file?.delete()
